# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BuildGraph AI Prototype API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
...
  /api/auth/login:
    post:
      summary: Email login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: JWT token and user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
  /api/users:
    post:
      summary: Sign up user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupRequest"
      responses:
        "201":
          description: Created user
  /api/auth/me:
    get:
      summary: Current user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
...
  /api/agent-logs/{id}:
    get:
      summary: Agent log detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Log summary
...
  /api/as-tickets/{id}:
    get:
      summary: AS ticket detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
    SignupRequest:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
...